abstract out general search rules
